server/requirements.txt
click==8.1.3
Flask==2.1.2
Flask-Cors==3.0.10
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.36
uWSGI==2.0.20
Werkzeug==2.1.2
zipp==3.8.0
